changed to getByRole for testing accessibility
