+++ client/src/Home.test.js
@@ -1,5 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
+import { fireEvent } from "@testing-library/react";
 import Home from "./Home";
 
 describe("Home component has rendered correctly", () => {
@@ -9,8 +10,8 @@ describe("Home component has rendered correctly", () => {
         <Home />
       </BrowserRouter>
     );
-    const inputElement_1 = screen.getByLabelText(/Zip Code/i);
-    expect(inputElement_1).toBeInTheDocument();
+    const inputElement = screen.getByRole("textbox", { name: /Zip Code/i });
+    expect(inputElement).toBeInTheDocument();
   });
 
   test("Input Element 'Country Code' to be in the document", () => {
@@ -19,7 +20,9 @@ describe("Home component has rendered correctly", () => {
         <Home />
       </BrowserRouter>
     );
-    const inputElement_2 = screen.getByLabelText(/Country Code/i);
+    const inputElement_2 = screen.getByRole("textbox", {
+      name: /Country Code/i,
+    });
     expect(inputElement_2).toBeInTheDocument();
   });
 
@@ -32,13 +35,34 @@ describe("Home component has rendered correctly", () => {
     const headingElement = screen.getByText("Weather Agent");
     expect(headingElement).toBeInTheDocument();
   });
-  test("Button Element to be in the document", () => {
+  test("Button Element 'Weather Report' to be in the document", () => {
     render(
       <BrowserRouter>
         <Home />
       </BrowserRouter>
     );
-    const buttonElement = screen.getByRole("button");
+    const buttonElement = screen.getByText(/Weather Report/i);
     expect(buttonElement).toBeInTheDocument();
+  });
+
+  it("should navigate to getweather/400601/IN upon providing a zip code and a country code", () => {
+    render(
+      <BrowserRouter>
+        <Home />
+      </BrowserRouter>
+    );
+    const inputZip = screen.getByLabelText(/Zip Code/i);
+    const inputCountry = screen.getByLabelText(/Country Code/i);
+    fireEvent.change(inputZip, {
+      target: { value: "400601" },
+    });
+    expect(inputZip).toHaveValue("400601");
+    fireEvent.change(inputCountry, {
+      target: { value: "IN" },
+    });
+    expect(inputCountry).toHaveValue("IN");
+    const buttonElement = screen.getByRole("button");
+    fireEvent.click(buttonElement);
+    expect(window.location.href).toBe("http://localhost/getweather/400601/IN");
   });
 });
